# Conflicts:
#	GameVersion.java
#	Maze.java
#	Maze2D.java
#	MazeInterface.java
#	Menu.java
#	MultiPlayerVersion.java
#	Player.java
#	SoloVersion.java
#	Test.java
#	TimeTrialVersion.java
#	View.java
#	bestSolos.txt
#	hostMenu.java
#	menu.css

+++ Player.java
@@ -30,8 +30,8 @@ public class Player implements Serializable{
 		//shape.setFill(Color.BLUE);
 	}
 
-	public double radius(){
-		return shape.getRadius();
+	public float radius(){
+		return (float)shape.getRadius();
 	}
 
 	public void setPosition(Point2D position, int ori){
